+++ Module.Persist.TPM/Model/TPM/COGS.cs
@@ -41,7 +41,7 @@ namespace Module.Persist.TPM.Model.TPM
         public int Year { get; set; }
         public int ClientTreeId { get; set; }
         public Guid? BrandTechId { get; set; }
-        public int Volume { get; set; }
+        public double Volume { get; set; }
         public bool Disabled { get; set; }
         public DateTimeOffset? DeletedDate { get; set; }
 
@@ -59,7 +59,7 @@ namespace Module.Persist.TPM.Model.TPM
         public int Year { get; set; }
         public int ClientTreeId { get; set; }
         public Guid? BrandTechId { get; set; }
-        public int Volume { get; set; }
+        public double Volume { get; set; }
         public bool Disabled { get; set; }
         public DateTimeOffset? DeletedDate { get; set; }
